Modularised CreateProperty.tsx into re-usable components
Modularised API calls

Reduced file size on CreateProperty.tsx
